requirements.txt
alembic==1.14.1
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==2.3.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
intasend-python==1.1.2
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==2.1.5
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
WTForms==3.1.2